src/tests/testapp/requirements.txt
mock==1.3.0
nose==1.3.7
yapsy==1.11.223
pyyaml==3.11
django==1.5.5
djangorestframework==2.3.8
six==1.10.0

#drf2swagger==1.0
#py2swagger==1.0